refactor(*): Move dev-time modules to scripts folder
