+++ package.json
@@ -38,12 +38,14 @@
   "author": "accordproject.org",
   "license": "Apache-2.0",
   "devDependencies": {
+    "acorn": "5.1.2",
     "babel-preset-env": "1.7.0",
     "babel-preset-latest": "6.24.1",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
     "chai-things": "0.2.0",
     "commander": "2.20.0",
+    "doctrine": "3.0.0",
     "eslint": "6.0.1",
     "fs-extra": "8.1.0",
     "jsdoc": "3.6.3",
@@ -56,6 +58,7 @@
     "npm-paths": "2.0.0",
     "nyc": "14.1.1",
     "pegjs": "0.10.0",
+    "semver": "6.2.0",
     "sinon": "7.3.2",
     "sinon-chai": "3.3.0",
     "webpack": "4.35.2",
@@ -63,16 +66,13 @@
     "yargs": "13.2.4"
   },
   "dependencies": {
-    "acorn": "5.1.2",
     "axios": "0.19.0",
     "debug": "4.1.1",
     "lorem-ipsum": "1.0.6",
     "mkdirp": "0.5.1",
     "moment-mini": "2.22.1",
-    "semver": "6.2.0",
     "urijs": "1.19.1",
-    "uuid": "3.3.2",
-    "doctrine": "3.0.0"
+    "uuid": "3.3.2"
   },
   "license-check-and-add-config": {
     "folder": "./lib",
@@ -123,8 +123,6 @@
       "index.js",
       "webpack.config.js",
       "lib/codegen/codegen.js",
-      "lib/codegen/parsejs.js",
-      "lib/codegen/javascriptparser.js",
       "lib/introspect/parser.js",
       "out/**",
       "scripts/**",
