+++ scripts/api-changelog.js
@@ -28,7 +28,7 @@ if (fs.existsSync(apiSignatureFile)) {
 return Promise.resolve()
     .then(() => {
         return new Promise((resolve, reject) => {
-            const parsejs = path.resolve(parentDirectory, 'scripts', 'parsejs.js');
+            const parsejs = path.resolve(parentDirectory, 'scripts', 'codegen', 'parsejs.js');
             const command = child_process.fork(parsejs, ['--format', 'APISignature', '--inputDir', path.resolve(parentDirectory, 'lib'), '--outputDir', parentDirectory]);
             command.on('exit', (code) => {
                 if (code !== 0) {
